+++ day16/rust/mr-kaffee/src/main.rs
@@ -4,9 +4,9 @@ use std::time::Instant;
 /// puzzle input
 pub const INPUT: &str = include_str!("../input.txt");
 /// expected result part 1
-pub const EXP_1: usize = 883;
+pub const EXP_1: u64 = 883;
 /// expected result part 2
-pub const EXP_2: usize = 1_675_198_555_015;
+pub const EXP_2: u64 = 1_675_198_555_015;
 
 fn main() {     
     let start = Instant::now();
